+++ crates/kvdb/src/kv_manager/error.rs
@@ -1,3 +1,18 @@
+// Copyright (C) 2023 Entropy Cryptography Inc.
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
 //! Custom error types for [crate::kv_manager].
 
 use crate::encrypted_sled;
@@ -36,4 +51,5 @@ pub enum InnerKvError {
     #[error("Deserialization Error: failed to deserialize kvstore bytes")]
     DeserializationErr,
 }
+
 pub(super) type InnerKvResult<Success> = Result<Success, InnerKvError>;