+++ crates/threshold-signature-server/src/user/api.rs
@@ -1,3 +1,18 @@
+// Copyright (C) 2023 Entropy Cryptography Inc.
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
 use std::{str::FromStr, sync::Arc, time::SystemTime};
 
 use axum::{
@@ -19,7 +34,9 @@ use entropy_kvdb::kv_manager::{
 };
 use entropy_protocol::SigningSessionInfo;
 use entropy_protocol::ValidatorInfo;
-use entropy_shared::{types::KeyVisibility, OcwMessageDkg, X25519PublicKey, SIGNING_PARTY_SIZE};
+use entropy_shared::{
+    types::KeyVisibility, HashingAlgorithm, OcwMessageDkg, X25519PublicKey, SIGNING_PARTY_SIZE,
+};
 use futures::{
     channel::mpsc,
     future::{join_all, FutureExt},
@@ -52,7 +69,7 @@ use crate::{
         substrate::{
             get_program, get_registered_details, get_subgroup, return_all_addresses_of_subgroup,
         },
-        user::{check_in_registration_group, do_dkg, send_key},
+        user::{check_in_registration_group, compute_hash, do_dkg, send_key},
         validator::get_signer,
     },
     signing_client::{ListenerState, ProtocolErr},
@@ -73,6 +90,8 @@ pub struct UserSignatureRequest {
     pub validators_info: Vec<ValidatorInfo>,
     /// When the message was created and signed
     pub timestamp: SystemTime,
+    /// Hashing algorithm to be used for signing
+    pub hash: HashingAlgorithm,
 }
 
 /// Type for validators to send user key's back and forth
@@ -127,14 +146,43 @@ pub async fn sign_tx(
 
     let message = hex::decode(&user_sig_req.message)?;
     let auxilary_data = user_sig_req.auxilary_data.as_ref().map(hex::decode).transpose()?;
-    let message_hash = Hasher::keccak(&message);
-    let message_hash_hex = hex::encode(message_hash);
-    let subgroup_signers = get_current_subgroup_signers(&api, &rpc, &message_hash_hex).await?;
+    
+    let mut runtime = Runtime::new();
+    let signature_request = SignatureRequest { message, auxilary_data };
+    
+    if user_details.program_pointers.0.is_empty() {
+        return Err(UserErr::NoProgramPointerDefined());
+    }
+    
+    for program_pointer in &user_details.program_pointers.0 {
+        let program = get_program(&api, &rpc, program_pointer).await?;
+        runtime.evaluate(&program, &signature_request)?;
+    }
+    let hashing_program = get_program(&api, &rpc, &user_details.program_pointers.0[0]).await?;
+    // We decided to do Keccak for subgroup selection for frontend compatability
+    let message_hash_keccak = compute_hash(
+        &HashingAlgorithm::Keccak,
+        &mut runtime,
+        hashing_program.as_slice(),
+        signature_request.message.as_slice(),
+    )?;
+    let message_hash_keccak_hex = hex::encode(message_hash_keccak);
+
+    let subgroup_signers =
+        get_current_subgroup_signers(&api, &rpc, &message_hash_keccak_hex).await?;
     check_signing_group(&subgroup_signers, &user_sig_req.validators_info, signer.account_id())?;
 
     // Use the validator info from chain as we can be sure it is in the correct order and the
     // details are correct
     user_sig_req.validators_info = subgroup_signers;
+
+    let message_hash = compute_hash(
+        &user_sig_req.hash,
+        &mut runtime,
+        program.as_slice(),
+        signature_request.message.as_slice(),
+    )?;
+    let message_hash_hex = hex::encode(message_hash);
 
     let signing_session_id = SigningSessionInfo {
         account_id: SubxtAccountId32(*signed_msg.account_id().as_ref()),
@@ -144,16 +192,6 @@ pub async fn sign_tx(
     let has_key = check_for_key(&signing_address, &app_state.kv_store).await?;
     if !has_key {
         recover_key(&api, &rpc, &app_state.kv_store, &signer, signing_address).await?
-    }
-
-    let mut runtime = Runtime::new();
-    let signature_request = SignatureRequest { message, auxilary_data };
-    if user_details.program_pointers.0.is_empty() {
-        return Err(UserErr::NoProgramPointerDefined());
-    }
-    for program_pointer in &user_details.program_pointers.0 {
-        let program = get_program(&api, &rpc, program_pointer).await?;
-        runtime.evaluate(&program, &signature_request)?;
     }
 
     let (mut response_tx, response_rx) = mpsc::channel(1);