+++ crates/threshold-signature-server/tests/protocol_wasm.rs
@@ -1,13 +1,31 @@
+// Copyright (C) 2023 Entropy Cryptography Inc.
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
 #![cfg(feature = "wasm_test")]
+
 //! Integration tests which use a nodejs process to test wasm bindings to the entropy-protocol
 //! client functions.
 //!
 //! These tests require additional build steps and are not run by default.
+
 mod helpers;
+
 use axum::http::StatusCode;
 use entropy_kvdb::clean_tests;
 use entropy_protocol::{KeyParams, ValidatorInfo};
-use entropy_shared::{KeyVisibility, OcwMessageDkg};
+use entropy_shared::{HashingAlgorithm, KeyVisibility, OcwMessageDkg};
 use entropy_testing_utils::{
     chain_api::entropy::runtime_types::bounded_collections::bounded_vec::BoundedVec,
     constants::{
@@ -56,7 +74,7 @@ async fn test_wasm_sign_tx_user_participates() {
     let one = AccountKeyring::Eve;
     let dave = AccountKeyring::Dave;
 
-    let signing_address = one.clone().to_account_id().to_ss58check();
+    let signing_address = one.to_account_id().to_ss58check();
     let (validator_ips, _validator_ids, users_keyshare_option) =
         spawn_testing_validators(Some(signing_address.clone()), true).await;
     let substrate_context = test_context_stationary().await;
@@ -92,6 +110,7 @@ async fn test_wasm_sign_tx_user_participates() {
         auxilary_data: Some(hex::encode(AUXILARY_DATA_SHOULD_SUCCEED)),
         validators_info: validators_info.clone(),
         timestamp: SystemTime::now(),
+        hash: HashingAlgorithm::Keccak,
     };
 
     let submit_transaction_requests =
@@ -141,7 +160,7 @@ async fn test_wasm_sign_tx_user_participates() {
     .await;
 
     // Check that the signature the user gets matches the first of the server's signatures
-    let user_sig = if let Some(user_sig_stripped) = user_sig.strip_suffix("\n") {
+    let user_sig = if let Some(user_sig_stripped) = user_sig.strip_suffix('\n') {
         user_sig_stripped.to_string()
     } else {
         user_sig